backend/pyproject.toml
[build-system]
requires = ["hatchling"]
build-backend = "hatchling.build"

[project]
name = "multi-source-analysis-agent"
version = "1.0.0"
description = "AI-powered research assistant that analyzes uploaded documents, fetches statistical data, and provides web context"
readme = "README.md"
license = {text = "MIT"}
authors = [
    {name = "AI Policy Analyzer Team"}
]
keywords = ["ai", "rag", "policy-analysis", "research", "langchain", "fastapi"]
classifiers = [
    "Development Status :: 4 - Beta",
    "Intended Audience :: Developers",
    "Intended Audience :: Science/Research",
    "License :: OSI Approved :: MIT License",
    "Programming Language :: Python :: 3",
    "Programming Language :: Python :: 3.10",
    "Programming Language :: Python :: 3.11",
    "Programming Language :: Python :: 3.12",
    "Topic :: Scientific/Engineering :: Artificial Intelligence",
    "Topic :: Internet :: WWW/HTTP :: HTTP Servers",
]
requires-python = ">=3.10"
dependencies = [
    "fastapi>=0.104.0",
    "uvicorn[standard]>=0.24.0",
    "python-multipart>=0.0.6",
    "langchain>=0.1.0",
    "langchain-openai>=0.0.2",
    "langchain-community>=0.0.10",
    "langchain-experimental>=0.0.47", 
    "qdrant-client>=1.7.0",
    "langchain-qdrant>=0.0.1",
    "pymupdf>=1.23.0",
    "pillow>=10.0.0",
    "python-dotenv>=1.0.0",
    "langsmith>=0.0.77",
    "tavily-python>=0.3.0",
    "pydantic>=2.5.0",
    "pydantic-settings>=2.1.0",
    "aiofiles>=23.2.0",
    "aiohttp>=3.8.0",  # For testing client
]

[project.optional-dependencies]
dev = [
    "pytest>=7.0.0",
    "pytest-asyncio>=0.21.0",
    "httpx>=0.24.0",
    "black>=23.0.0",
    "isort>=5.12.0",
    "flake8>=6.0.0",
    "mypy>=1.0.0",
]
test = [
    "pytest>=7.0.0",
    "pytest-asyncio>=0.21.0",
    "httpx>=0.24.0",
    "pytest-cov>=4.0.0",
]
datacommons = [
    # Real Data Commons integration via data_gemma (requires Python 3.10+)
    "data-gemma @ git+https://github.com/datacommonsorg/llm-tools.git",
    # Required Google Cloud dependencies for data_gemma
    "google-cloud-aiplatform>=1.34.0",
    "google-auth>=2.17.0",
    "transformers>=4.21.0",
    "torch>=1.12.0",
]

[project.urls]
"Homepage" = "https://github.com/your-org/multi-source-analysis-agent"
"Bug Reports" = "https://github.com/your-org/multi-source-analysis-agent/issues"
"Source" = "https://github.com/your-org/multi-source-analysis-agent"
"Documentation" = "https://github.com/your-org/multi-source-analysis-agent#readme"

[project.scripts]
msaa-server = "main:main"

[tool.hatch.build.targets.wheel]
packages = ["utils", "tools"]
include = [
    "*.py",
    "utils/**/*.py",
    "tools/**/*.py",
]

[tool.hatch.metadata]
allow-direct-references = true

[tool.black]
line-length = 100
target-version = ['py310', 'py311', 'py312']
include = '\.pyi?$'
extend-exclude = '''
/(
  # directories
  \.eggs
  | \.git
  | \.hg
  | \.mypy_cache
  | \.tox
  | \.venv
  | build
  | dist
)/
'''

[tool.isort]
profile = "black"
line_length = 100
multi_line_output = 3
include_trailing_comma = true
force_grid_wrap = 0
use_parentheses = true
ensure_newline_before_comments = true

[tool.mypy]
python_version = "3.10"
warn_return_any = true
warn_unused_configs = true
disallow_untyped_defs = true
disallow_incomplete_defs = true
check_untyped_defs = true
disallow_untyped_decorators = true
no_implicit_optional = true
warn_redundant_casts = true
warn_unused_ignores = true
warn_no_return = true
warn_unreachable = true
strict_equality = true

[tool.pytest.ini_options]
testpaths = ["tests"]
python_files = ["test_*.py", "*_test.py"]
python_classes = ["Test*"]
python_functions = ["test_*"]
addopts = "-v --tb=short --strict-markers"
markers = [
    "slow: marks tests as slow (deselect with '-m \"not slow\"')",
    "integration: marks tests as integration tests",
    "unit: marks tests as unit tests",
]
asyncio_mode = "auto" 